refactor mailers
