+++ app/jobs/rsvp_job.rb
@@ -2,6 +2,6 @@ class RsvpJob
   @queue = :rsvp
   
   def self.perform(event_id, user_id)
-    RsvpMailer.confirmation(event_id, user_id).deliver
+    WormholeMailer.rsvp_confirmation(event_id, user_id).deliver
   end
 end
