+++ app/views/wormhole_mailer/event_cancellation.html.erb
