+++ app/views/wormhole_mailer/new_post.html.erb
